change types.rs to file_wallet
