+++ wallet/src/client.rs
@@ -23,7 +23,6 @@ use tokio_core::reactor;
 
 use libwallet::types::*;
 use std::io;
-use types::*;
 use util::LOGGER;
 
 /// Call the wallet API to create a coinbase output for the given block_fees.
