+++ wallet/src/info.rs
@@ -17,7 +17,7 @@ use core::core::amount_to_hr_string;
 use keychain::Keychain;
 use libwallet::types::*;
 use prettytable;
-use types::*;
+use file_wallet::*;
 
 pub fn show_info(config: &WalletConfig, keychain: &Keychain) {
 	let wallet_info = retrieve_info(config, keychain);
