+++ wallet/src/libwallet/types.rs
@@ -276,6 +276,58 @@ impl<'de> serde::de::Visitor<'de> for BlockIdentifierVisitor {
 	}
 }
 
+/// Amount in request to build a coinbase output.
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub enum WalletReceiveRequest {
+	Coinbase(BlockFees),
+	PartialTransaction(String),
+	Finalize(String),
+}
+
+/// Fees in block to use for coinbase amount calculation
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct BlockFees {
+	pub fees: u64,
+	pub height: u64,
+	pub key_id: Option<Identifier>,
+}
+
+impl BlockFees {
+	pub fn key_id(&self) -> Option<Identifier> {
+		self.key_id.clone()
+	}
+}
+
+/// Response to build a coinbase output.
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct CbData {
+	pub output: String,
+	pub kernel: String,
+	pub key_id: String,
+}
+
+/// a contained wallet info struct, so automated tests can parse wallet info
+/// can add more fields here over time as needed
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct WalletInfo {
+	// height from which info was taken
+	pub current_height: u64,
+	// total amount in the wallet
+	pub total: u64,
+	// amount awaiting confirmation
+	pub amount_awaiting_confirmation: u64,
+	// confirmed but locked
+	pub amount_confirmed_but_locked: u64,
+	// amount currently spendable
+	pub amount_currently_spendable: u64,
+	// amount locked by previous transactions
+	pub amount_locked: u64,
+	// whether the data was confirmed against a live node
+	pub data_confirmed: bool,
+	// node confirming the data
+	pub data_confirmed_from: String,
+}
+
 #[derive(Debug)]
 pub struct Error {
 	inner: Context<ErrorKind>,
