+++ wallet/src/receiver.rs
@@ -31,7 +31,7 @@ use keychain::Keychain;
 use libtx::{reward, slate::Slate};
 use libwallet::types::*;
 use libwallet::{keys, selection};
-use types::*;
+use file_wallet::*;
 use util::LOGGER;
 
 /// Dummy wrapper for the hex-encoded serialized transaction.
