+++ wallet/src/server.rs
@@ -16,7 +16,7 @@ use api::ApiServer;
 use handlers::CoinbaseHandler;
 use keychain::Keychain;
 use receiver::WalletReceiver;
-use types::WalletConfig;
+use file_wallet::WalletConfig;
 use util::LOGGER;
 
 pub fn start_rest_apis(wallet_config: WalletConfig, keychain: Keychain) {
